BankNote_test/requirements.txt
torch
torchvision
Pillow
numpy
matplotlib
pandas